+++ src/main/java/de/maxhenkel/pipez/integration/waila/PluginPipes.java
@@ -1,7 +1,9 @@
 package de.maxhenkel.pipez.integration.waila;
 
+import de.maxhenkel.pipez.blocks.tileentity.UpgradeTileEntity;
 import mcp.mobius.waila.api.IRegistrar;
 import mcp.mobius.waila.api.IWailaPlugin;
+import mcp.mobius.waila.api.TooltipPosition;
 import mcp.mobius.waila.api.WailaPlugin;
 import net.minecraft.util.ResourceLocation;
 
@@ -14,6 +16,7 @@ public class PluginPipes implements IWailaPlugin {
 
     @Override
     public void register(IRegistrar registrar) {
-
+        registrar.registerComponentProvider(HUDHandlerPipes.INSTANCE, TooltipPosition.BODY, UpgradeTileEntity.class);
+        registrar.registerBlockDataProvider(HUDHandlerPipes.INSTANCE, UpgradeTileEntity.class);
     }
 }